Remove check_previous in setup.py
This check has a dependency on an internal pip interface that was
removed in pip 10, therefore our build no longer works using the
CircleCI Python 3.6 image.

This code was removed in [AIRFLOW-1582], and we don't need it anyway as
we're using a containerised Airflow setup, so there's no in-place
upgrades.